add-a-cat-to-that/requirements.txt
# Copyright 2019 Google, Inc.
#
#  Licensed under the Apache License, Version 2.0 (the "License");
#  you may not use this file except in compliance with the License.
#  You may obtain a copy of the License at
#
#      http://www.apache.org/licenses/LICENSE-2.0
#
#  Unless required by applicable law or agreed to in writing, software
#  distributed under the License is distributed on an "AS IS" BASIS,
#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
#  See the License for the specific language governing permissions and
#  limitations under the License.

asn1crypto==0.24.0
cbor2==4.1.2
cffi==1.12.3
Click==7.0
cryptography==2.7
Flask==1.0.3
future==0.17.1
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
Pillow==6.0.0
pycparser==2.19
pyOpenSSL==19.0.0
six==1.12.0
webauthn==0.4.5
Werkzeug==0.15.4
